\name{plvs-package}
\alias{plvs-package}
\alias{plvs}
\Rdversion{1.0}
\docType{package}
\title{
Ultra-high dimensional variable selection for piecewise linear loss function
}
\description{
This package selects the important variables for piecewise linear loss function under ultra-high dimensional data, and simultaneously
estimate the corresponding parameters of the selected variables, in which the Coordinate Descend and Minorization and Maximization (CDMM) algorithm is used.
}
\details{
\tabular{ll}{
Package: \tab plvs\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-07-03\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab Yes\cr
}
Unpenalized Composite Quantile Regression. cqr(x, y, q)

Penalized Composite Quantile Regression. pcqr(x, y, q)

Penalized Single Quantile Regression. pqr(x, y, q)
}
\author{
Xu Liu, Hongmei Jiang and Xingjie Shi

Maintainer: Xu Liu <liu.xu@mail.shufe.edu.cn>
}
\references{
Ultra-high dimensional variable selection piecewise linear loss function
}
%\keyword{ Quantile regression, LASSO, MCP, SCAD, Piecewise Linear, CDMM}
\seealso{cqr, pqr, pcqr, plot.cv, setuplambda}

